LayoutTests/imported/w3c/web-platform-tests/css/css-view-transitions/navigation/pageswap-traverse-navigation-no-bfcache.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "https://localhost:9443/css/css-view-transitions/navigation/pageswap-traverse-navigation-no-...

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "https://localhost:9443/css/css-view-transitions/navigation/pageswap-traverse-navigation-no-bfcache.https.html?popup", "traverse", "from", "pagehide"] length 6, got ["pagehide"] length 1

TIMEOUT pageswap on traverse navigation from script Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_array_equals: incorrect event order lengths differ, expected array ["pageswap", "transition", "https://localhost:9443/css/css-view-transitions/navigation/pageswap-traverse-navigation-no-bfcache.https.html?popup", "traverse", "from", "pagehide"] length 6, got ["pagehide"] length 1

TIMEOUT pageswap on traverse navigation from script Test timed out
